tools/requirements.txt
opencv-python>=4.9.0
numpy>=1.26.0
pytesseract>=0.3.13
python-dotenv>=1.0.0
openai>=1.30.0
